+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Pashmak Changelog
 
+## 0.4 (2020-12-1)
+
+- added new internal module `test` contains assertion functions
+- added a feature to module path system to import directories as `__init__.pashm` of them
+- fixed pashamk cli signal handling
+
 ## 0.3.1 (2020-11-27)
 
 - fixed syntax removing more than 1 space bug. now 'println "hello&nbsp;&nbsp;&nbsp;&nbsp;world"' will print "hello&nbsp;&nbsp;&nbsp;&nbsp;world", not `hello world`